+++ src/main/java/mll/service/SubmissionService.java
@@ -401,7 +401,7 @@ public class SubmissionService
 			owner.setPrimaryPhone((String)writer.get("primaryPhone"));
 			owner.setSecondaryPhone((String)writer.get("secondaryPhone"));
 			owner.setContribution((String)writer.get("contribution"));
-			owner.setOwner_percent((Integer)writer.get("ownershipPercent"));
+			owner.setOwner_percent((Long)writer.get("ownershipPercent"));
 			owner.setRole((String)writer.get("MusicianRole"));
 	    	
 	    	metadata.getOwners().add(owner);
@@ -444,7 +444,7 @@ public class SubmissionService
 			owner.setPrimaryPhone((String)recorder.get("primaryPhone"));
 			owner.setSecondaryPhone((String)recorder.get("secondaryPhone"));
 			owner.setContribution((String)recorder.get("contribution"));
-			owner.setOwner_percent((Integer)recorder.get("ownershipPercent"));
+			owner.setOwner_percent((Long)recorder.get("ownershipPercent"));
 			owner.setRole((String)recorder.get("MusicianRole"));
 			
 	    	metadata.getOwners().add(owner);
